[Cleanup] Moved chat-connected dto interfaces to `src/common/types/chats.types.ts`
